+++ src/index.ts
@@ -4,4 +4,6 @@ export { DEFAULT_GITHUB_REGISTRY } from './consts.js';
 export { BaseRegistry, CHAIN_FILE_REGEX } from './registry/BaseRegistry.js';
 export { GithubRegistry, GithubRegistryOptions } from './registry/GithubRegistry.js';
 export { ChainFiles, IRegistry, RegistryContent, RegistryType } from './registry/IRegistry.js';
+export { MergedRegistry, MergedRegistryOptions } from './registry/MergedRegistry.js';
+export { PartialRegistry, PartialRegistryOptions } from './registry/PartialRegistry.js';
 export { ChainAddresses, ChainAddressesSchema } from './types.js';