+++ src/registry/IRegistry.ts
@@ -4,6 +4,7 @@ import type { ChainAddresses, MaybePromise } from '../types.js';
 export interface ChainFiles {
   metadata?: string;
   addresses?: string;
+  logo?: string;
 }
 
 export interface RegistryContent {
@@ -15,7 +16,9 @@ export interface RegistryContent {
 
 export enum RegistryType {
   Github = 'github',
-  Local = 'local',
+  FileSystem = 'filesystem',
+  Merged = 'merged',
+  Partial = 'partial',
 }
 
 export interface IRegistry {
@@ -25,10 +28,15 @@ export interface IRegistry {
   listRegistryContent(): MaybePromise<RegistryContent>;
 
   getChains(): MaybePromise<Array<ChainName>>;
+
   getMetadata(): MaybePromise<ChainMap<ChainMetadata>>;
   getChainMetadata(chainName: ChainName): MaybePromise<ChainMetadata | null>;
+
   getAddresses(): MaybePromise<ChainMap<ChainAddresses>>;
   getChainAddresses(chainName: ChainName): MaybePromise<ChainAddresses | null>;
+
+  getChainLogoUri(chainName: ChainName): Promise<string | null>;
+
   addChain(chain: {
     chainName: ChainName;
     metadata?: ChainMetadata;
@@ -43,4 +51,6 @@ export interface IRegistry {
 
   addWarpRoute(config: WarpCoreConfig): MaybePromise<void>;
   // TODO define more deployment artifact related methods
+
+  merge(otherRegistry: IRegistry): IRegistry;
 }