+++ test/unit/registry.test.ts
@@ -4,13 +4,18 @@ import { expect } from 'chai';
 import type { ChainMetadata } from '@hyperlane-xyz/sdk';
 import fs from 'fs';
 import { CHAIN_FILE_REGEX } from '../../src/registry/BaseRegistry.js';
+import { FileSystemRegistry } from '../../src/registry/FileSystemRegistry.js';
 import { GithubRegistry } from '../../src/registry/GithubRegistry.js';
-import { LocalRegistry } from '../../src/registry/LocalRegistry.js';
+import { RegistryType } from '../../src/registry/IRegistry.js';
+import { MergedRegistry } from '../../src/registry/MergedRegistry.js';
+import { PartialRegistry } from '../../src/registry/PartialRegistry.js';
 import { ChainAddresses } from '../../src/types.js';
 
 const MOCK_CHAIN_NAME = 'mockchain';
 const MOCK_CHAIN_NAME2 = 'mockchain2';
+const MOCK_DISPLAY_NAME = 'faketherum';
 const MOCK_SYMBOL = 'MOCK';
+const MOCK_ADDRESS = '0x0000000000000000000000000000000000000001';
 
 describe('Registry utilities', () => {
   const githubRegistry = new GithubRegistry();
@@ -18,10 +23,19 @@ describe('Registry utilities', () => {
   expect(githubRegistry.repoName).to.eql('hyperlane-registry');
   expect(githubRegistry.branch).to.eql('main');
 
-  const localRegistry = new LocalRegistry({ uri: './' });
+  const localRegistry = new FileSystemRegistry({ uri: './' });
   expect(localRegistry.uri).to.be.a('string');
 
-  for (const registry of [githubRegistry, localRegistry]) {
+  const partialRegistry = new PartialRegistry({
+    chainMetadata: { ethereum: { chainId: 1, displayName: MOCK_DISPLAY_NAME } },
+    chainAddresses: { ethereum: { mailbox: MOCK_ADDRESS } },
+  });
+
+  const mergedRegistry = new MergedRegistry({
+    registries: [githubRegistry, localRegistry, partialRegistry],
+  });
+
+  for (const registry of [githubRegistry, localRegistry, partialRegistry, mergedRegistry]) {
     it(`Lists all chains for ${registry.type} registry`, async () => {
       const chains = await registry.getChains();
       expect(chains.length).to.be.greaterThan(0);
@@ -32,11 +46,16 @@ describe('Registry utilities', () => {
       const metadata = await registry.getMetadata();
       expect(Object.keys(metadata).length).to.be.greaterThan(0);
       expect(metadata['ethereum'].chainId).to.eql(1);
+      if (registry.type === RegistryType.Partial || registry.type === RegistryType.Merged) {
+        expect(metadata['ethereum'].displayName).to.eql(MOCK_DISPLAY_NAME);
+      } else {
+        expect(metadata['ethereum'].displayName).to.eql('Ethereum');
+      }
     }).timeout(10_000);
 
     it(`Fetches single chain metadata for ${registry.type} registry`, async () => {
       const metadata = await registry.getChainMetadata('ethereum');
-      expect(metadata.chainId).to.eql(1);
+      expect(metadata!.chainId).to.eql(1);
     }).timeout(5_000);
 
     it(`Fetches chain addresses for ${registry.type} registry`, async () => {
@@ -47,7 +66,7 @@ describe('Registry utilities', () => {
 
     it(`Fetches single chain addresses for ${registry.type} registry`, async () => {
       const addresses = await registry.getChainAddresses('ethereum');
-      expect(addresses.mailbox.substring(0, 2)).to.eql('0x');
+      expect(addresses!.mailbox.substring(0, 2)).to.eql('0x');
     }).timeout(5_000);
 
     it(`Caches correctly for ${registry.type} registry`, async () => {
@@ -59,14 +78,14 @@ describe('Registry utilities', () => {
     }).timeout(250);
 
     // TODO remove this once GitHubRegistry methods are implemented
-    if (registry.type === 'github') continue;
+    if (registry.type !== RegistryType.FileSystem) continue;
 
     it(`Adds a new chain for ${registry.type} registry`, async () => {
       const mockMetadata: ChainMetadata = {
-        ...(await registry.getChainMetadata('ethereum')),
+        ...(await registry.getChainMetadata('ethereum'))!,
         name: MOCK_CHAIN_NAME,
       };
-      const mockAddresses: ChainAddresses = await registry.getChainAddresses('ethereum');
+      const mockAddresses: ChainAddresses = await registry.getChainAddresses('ethereum')!;
       await registry.addChain({
         chainName: MOCK_CHAIN_NAME,
         metadata: mockMetadata,
@@ -105,5 +124,7 @@ describe('Registry regex', () => {
     expect(CHAIN_FILE_REGEX.test('chains/ethereum/metadata.yaml')).to.be.true;
     expect(CHAIN_FILE_REGEX.test('chains/ancient8/addresses.yaml')).to.be.true;
     expect(CHAIN_FILE_REGEX.test('chains/_NotAChain/addresses.yaml')).to.be.false;
+    expect(CHAIN_FILE_REGEX.test('chains/foobar/logo.svg')).to.be.true;
+    expect(CHAIN_FILE_REGEX.test('chains/foobar/randomfile.txt')).to.be.false;
   });
 });