+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hyperlane-xyz/registry
 
+## 1.2.0
+
+### Minor Changes
+
+- 3d8a8d0: Added holesky chain metadata and core artifacts
+
+## 1.1.2
+
+### Patch Changes
+
+- 46b136f: Add getChainLogoUri method and improve caching
+
 ## 1.1.1
 
 ### Patch Changes